+++ streammq-spring-cloud-stream-binder/src/test/java/io/github/streammq/spring/cloud/stream/binder/StreamMQMessageBinderTest.java
@@ -8,8 +8,6 @@ package io.github.streammq.spring.cloud.stream.binder;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyInt;
-import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;